Fixed OpenCL build failure after b302561b763a [SVE][IR] Scalable Vector size queries and IR instruction support.
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214128
Tested-by: Tianqing Wang <tianqing.wang@intel.com>

+++ opencl-intel/backend/vectorizer/Packetizer/Packetizer.cpp
@@ -52,7 +52,7 @@ EnableScatterGather_v4i8("gather-scatter-v4i8", cl::init(false), cl::Hidden,
 static unsigned generateExplicitAlignment(unsigned Alignment, const PointerType* PT) {
   if (!Alignment)
     Alignment = std::max(
-        1U,
+        uint64_t(1U),
         PT->getElementType()->getPrimitiveSizeInBits().getKnownMinSize() / 8);
   return Alignment;
 }
